refactor(vsock-host): split connection core
