+++ crates/vsock-host/src/lib.rs
@@ -67,35 +67,28 @@
 //! [`ExecControlOutcome`] so callers can distinguish delivered requests from
 //! guest statuses and guest error responses.
 
+mod connection;
 mod exec_operation;
 mod file;
 mod operation_tracker;
 #[cfg(test)]
 mod tests;
 
-use std::collections::HashMap;
 use std::io;
 use std::os::unix::io::RawFd;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU32, Ordering};
 use std::time::Duration;
 
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{UnixListener, UnixStream};
-use tokio::sync::{Notify, oneshot};
 use tokio::task::JoinHandle;
-use tokio::time::{self, Instant};
 
-use operation_tracker::{
-    NormalOperationFenceRejection as TrackerNormalOperationFenceRejection,
-    NormalOperationRejection, NormalOperationToken, NormalOperationTracker,
-    NormalOperationTransitionError, NormalOperationTransitionHandle,
+use connection::{
+    CompositeNormalOperation, ConnectionState, Shared, normal_operation_transition_error,
+    normal_request_on_shared_with_write_observer_frame_builder,
+    request_on_shared_with_composite_operation_and_observer_frame_builder,
 };
-use vsock_proto::{
-    BorrowedRawMessage, DecodeWithError, Decoder, MSG_ERROR, MSG_OPERATIONS_QUIESCED,
-    MSG_OPERATIONS_RESUMED, MSG_PING, MSG_PONG, MSG_QUIESCE_OPERATIONS, MSG_READY,
-    MSG_RESUME_OPERATIONS, MSG_SHUTDOWN, MSG_SHUTDOWN_ACK, RawMessage,
-};
+#[cfg(test)]
+use connection::{RequestWriteGuard, write_request_frame_with_builder};
+use operation_tracker::NormalOperationFenceRejection as TrackerNormalOperationFenceRejection;
 
 pub use exec_operation::{
     ExecCaptureRequest, ExecControlAck, ExecControlGuestStatus, ExecControlHandle,
@@ -104,8 +97,6 @@ pub use exec_operation::{
     SupervisedExecControl, SupervisedExecHandle, SupervisedExecRequest,
 };
 pub use file::{CopyFileOptions, CopyFileResult, WriteFileEntry};
-
-const READ_BUF_SIZE: usize = 64 * 1024;
 
 /// Observer called when a request frame reaches the guest-write boundary.
 ///
@@ -221,260 +212,6 @@ impl std::error::Error for FencedExecError {
     }
 }
 
-/// Connection lifecycle, expressed as data rather than a separate atomic flag.
-///
-/// Pending requests, active exec operations, and connected process state
-/// live inside the `Connected` variant so registrations are structurally
-/// unreachable once the reader task has exited.
-///
-/// The invariant "connection is closed ⇔ registrations are impossible" is
-/// enforced by the type: every code path that cares about liveness must
-/// `match`, which precludes the old footgun of reading a stale close flag
-/// without taking the corresponding lock.
-enum ConnectionState {
-    Connected {
-        /// Pending request responses: seq → response route.
-        pending: HashMap<u32, PendingResponse>,
-        /// Active exec-operation state owned by the exec_operation module.
-        operations: exec_operation::Operations,
-    },
-    Closed,
-}
-
-/// Shared state between the reader task and public API methods.
-struct Shared {
-    /// Serialises writes to the stream.
-    writer: tokio::sync::Mutex<tokio::net::unix::OwnedWriteHalf>,
-    /// Serialises memory-heavy encoded frame construction without blocking
-    /// ordinary writer-lock users such as lifecycle and control frames.
-    frame_builder: tokio::sync::Mutex<()>,
-    /// Serialises file-write request lifecycles through their terminal
-    /// responses while allowing control and lifecycle frames to bypass them.
-    file_write_gate: tokio::sync::Mutex<()>,
-    /// Raw fd of the underlying socket, used to poison a corrupted stream.
-    fd: RawFd,
-    /// Monotonically increasing sequence number (starts at 2, skips 0).
-    /// Handshake uses seq=1 before Shared is created, so post-handshake
-    /// sequences start at 2 to avoid collisions.
-    seq: AtomicU32,
-    /// Single source of truth for connection liveness plus all per-connection
-    /// registration tables. See [`ConnectionState`].
-    state: std::sync::Mutex<ConnectionState>,
-    /// Connection-local tracker for logical normal guest operations.
-    ///
-    /// Routing maps stay inside [`ConnectionState`]; this tracker records the
-    /// neutral operation-readiness facts that sandbox policy can consume later.
-    normal_operations: NormalOperationTracker,
-    /// Notified when the connection closes. Pure signalling — all state is in
-    /// `state`.
-    close_notify: Notify,
-    /// One-shot test hook at the unlocked exec-output payload-copy boundary.
-    #[cfg(test)]
-    exec_output_before_copy_hook: std::sync::Mutex<Option<Box<dyn FnOnce() + Send>>>,
-}
-
-struct ListenerSocketGuard {
-    path: Option<String>,
-}
-
-impl ListenerSocketGuard {
-    fn remove(&mut self) {
-        if let Some(path) = self.path.take() {
-            let _ = std::fs::remove_file(path);
-        }
-    }
-}
-
-impl Drop for ListenerSocketGuard {
-    fn drop(&mut self) {
-        self.remove();
-    }
-}
-
-struct PendingRequestGuard {
-    shared: Arc<Shared>,
-    seq: u32,
-}
-
-struct PendingResponse {
-    response_tx: oneshot::Sender<RawMessage>,
-    normal_operation: Option<PendingNormalOperation>,
-    normal_terminal_msg_types: &'static [u8],
-}
-
-enum PendingNormalOperation {
-    Owned(NormalOperationToken),
-    Composite(NormalOperationTransitionHandle),
-}
-
-struct RequestWriteGuard {
-    shared: Arc<Shared>,
-    write_started: bool,
-    write_returned: bool,
-}
-
-impl PendingRequestGuard {
-    fn new(shared: Arc<Shared>, seq: u32) -> Self {
-        Self { shared, seq }
-    }
-}
-
-impl Drop for PendingRequestGuard {
-    fn drop(&mut self) {
-        self.shared.remove_pending(self.seq);
-    }
-}
-
-impl RequestWriteGuard {
-    fn new(shared: Arc<Shared>) -> Self {
-        Self {
-            shared,
-            write_started: false,
-            write_returned: false,
-        }
-    }
-
-    fn mark_started(&mut self) {
-        self.write_started = true;
-    }
-
-    fn mark_returned(&mut self) {
-        self.write_returned = true;
-    }
-}
-
-impl Drop for RequestWriteGuard {
-    fn drop(&mut self) {
-        if self.write_started && !self.write_returned {
-            self.shared.poison_connection();
-        }
-    }
-}
-
-impl Shared {
-    /// Get next sequence number, skipping 0 (reserved for unsolicited messages).
-    fn next_seq(&self) -> u32 {
-        loop {
-            let seq = self.seq.fetch_add(1, Ordering::Relaxed);
-            if seq != 0 {
-                return seq;
-            }
-            // Wrapped to 0 — skip it.
-        }
-    }
-
-    /// Transition `Connected → Closed`, dropping registration maps outside the
-    /// state lock so sender drops (which wake their receivers) run without the
-    /// lock held. Idempotent: a second call leaves the existing closed state
-    /// untouched.
-    ///
-    /// `mem::replace` writes a placeholder `Closed` state so the old variant
-    /// can be moved out for destructuring. The already-`Closed` arm uses an `@`
-    /// binding to write the whole variant back unchanged, so double-close is a
-    /// structural no-op rather than a convention.
-    fn close(&self) {
-        self.close_with_reason("connection closed", ConnectionCloseKind::Closed);
-    }
-
-    fn close_with_reason(&self, reason: &'static str, kind: ConnectionCloseKind) {
-        let maps_to_drop = {
-            let mut guard = self.state.lock().unwrap_or_else(|e| e.into_inner());
-            match std::mem::replace(&mut *guard, ConnectionState::Closed) {
-                ConnectionState::Connected {
-                    pending,
-                    operations,
-                } => {
-                    // Serialize tracker close/poison with terminal dispatch,
-                    // which completes tracker tokens under this same state lock.
-                    match kind {
-                        ConnectionCloseKind::Closed => self.normal_operations.mark_closed(),
-                        ConnectionCloseKind::Poisoned => self.normal_operations.mark_not_parkable(),
-                    }
-                    let exec_operation_snapshot = operations.close_snapshot();
-                    *guard = ConnectionState::Closed;
-                    Some((pending, operations, exec_operation_snapshot))
-                }
-                closed @ ConnectionState::Closed => {
-                    // Reassign the whole variant by binding rather than by
-                    // convention.
-                    *guard = closed;
-                    None
-                }
-            }
-        };
-        if let Some((pending, operations, exec_operation_snapshot)) = maps_to_drop {
-            let maps = (pending, operations);
-            drop(maps);
-            self.close_notify.notify_waiters();
-            exec_operation::log_operations_closed(reason, &exec_operation_snapshot);
-        }
-    }
-
-    fn poison_connection(&self) {
-        self.close_with_reason("connection poisoned", ConnectionCloseKind::Poisoned);
-        let _ = nix::sys::socket::shutdown(self.fd, nix::sys::socket::Shutdown::Both);
-    }
-
-    fn remove_pending(&self, seq: u32) {
-        let mut guard = self.state.lock().unwrap_or_else(|e| e.into_inner());
-        if let ConnectionState::Connected { pending, .. } = &mut *guard {
-            pending.remove(&seq);
-        }
-    }
-
-    fn remove_operation(&self, seq: u32) {
-        let mut guard = self.state.lock().unwrap_or_else(|e| e.into_inner());
-        if let ConnectionState::Connected { operations, .. } = &mut *guard {
-            operations.remove(seq);
-        }
-    }
-
-    fn reserve_normal_operation(&self) -> io::Result<NormalOperationToken> {
-        self.normal_operations
-            .reserve()
-            .map_err(normal_operation_rejection_error)
-    }
-}
-
-pub(crate) struct CompositeNormalOperation {
-    normal_operation: Option<NormalOperationToken>,
-}
-
-impl CompositeNormalOperation {
-    fn reserve(shared: &Arc<Shared>) -> io::Result<Self> {
-        Ok(Self {
-            normal_operation: Some(shared.reserve_normal_operation()?),
-        })
-    }
-
-    fn transition_handle(&self) -> io::Result<NormalOperationTransitionHandle> {
-        self.normal_operation
-            .as_ref()
-            .map(NormalOperationToken::transition_handle)
-            .ok_or_else(|| {
-                io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    "composite normal operation already completed",
-                )
-            })
-    }
-
-    fn complete(mut self) -> io::Result<()> {
-        let normal_operation = self.normal_operation.take().ok_or_else(|| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                "composite normal operation already completed",
-            )
-        })?;
-        // Terminal proof can race with connection close, which clears tracker
-        // operations. Completion is idempotent for the composite owner.
-        match normal_operation.complete() {
-            Ok(()) | Err(NormalOperationTransitionError::UnknownOperation { .. }) => Ok(()),
-            Err(error) => Err(normal_operation_transition_error(error)),
-        }
-    }
-}
-
 struct FencedNormalOperationGuard {
     normal_operation: Option<CompositeNormalOperation>,
     fence: Option<NormalOperationFence>,
@@ -508,12 +245,6 @@ impl Drop for FencedNormalOperationGuard {
     }
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum ConnectionCloseKind {
-    Closed,
-    Poisoned,
-}
-
 /// Host-side vsock endpoint.
 ///
 /// Maintains a persistent connection to the guest agent and provides
@@ -536,604 +267,7 @@ pub struct VsockHost {
     fd: RawFd,
 }
 
-impl Drop for VsockHost {
-    fn drop(&mut self) {
-        // Drop registration state synchronously. The shutdown below normally
-        // lets `reader_loop` observe EOF and call `close()`, but aborting the
-        // reader task can win that race. Closing here makes active exec
-        // handles and stream receivers release immediately when the host is
-        // dropped.
-        self.shared.close();
-        // Signal EOF on the socket so the reader_loop's `read()` and the
-        // remote peer's blocking `read()` return immediately. Without this,
-        // the split stream halves keep the fd alive until the reader task is
-        // cancelled, which requires an async yield — not possible in Drop.
-        let _ = nix::sys::socket::shutdown(self.fd, nix::sys::socket::Shutdown::Both);
-        self._reader.abort();
-    }
-}
-
-/// Background reader task: owns the read half and decoder exclusively.
-///
-/// Dispatches responses and exec operation lifecycle frames by seq number.
-async fn reader_loop(
-    mut reader: tokio::net::unix::OwnedReadHalf,
-    mut decoder: Decoder,
-    shared: Arc<Shared>,
-) {
-    let mut buf = Box::new([0u8; READ_BUF_SIZE]);
-    loop {
-        let n = match reader.read(buf.as_mut()).await {
-            Ok(0) | Err(_) => break,
-            Ok(n) => n,
-        };
-        // n <= READ_BUF_SIZE guaranteed by read()
-        match decoder.decode_with(buf.get(..n).unwrap_or_default(), |msg| {
-            dispatch_reader_frame(&shared, msg)
-        }) {
-            Ok(()) => {}
-            Err(DecodeWithError::Protocol(_)) => break,
-            Err(DecodeWithError::Visitor(_)) => {
-                shared.poison_connection();
-                return;
-            }
-        }
-    }
-    // Connection lost — transition state to Closed. `close()` drops all
-    // registration maps outside the lock (waking every pending receiver
-    // with `RecvError`) and fires `close_notify` so test helpers wake.
-    shared.close();
-}
-
-fn dispatch_reader_frame(shared: &Arc<Shared>, msg: BorrowedRawMessage<'_>) -> io::Result<()> {
-    match exec_operation::dispatch_incoming_frame(shared, msg) {
-        Ok(true) => {
-            return Ok(());
-        }
-        Ok(false) => {}
-        Err(error) => {
-            return Err(error);
-        }
-    }
-
-    let pending_response = {
-        let mut guard = shared.state.lock().unwrap_or_else(|e| e.into_inner());
-        match &mut *guard {
-            ConnectionState::Connected { pending, .. } => {
-                if let Some(mut pending_response) = pending.remove(&msg.seq) {
-                    if pending_response
-                        .normal_terminal_msg_types
-                        .contains(&msg.msg_type)
-                        && let Some(normal_operation) = pending_response.normal_operation.take()
-                    {
-                        complete_pending_normal_operation(normal_operation)?;
-                    }
-                    Some(pending_response)
-                } else {
-                    None
-                }
-            }
-            ConnectionState::Closed => None,
-        }
-    };
-    if let Some(pending_response) = pending_response {
-        let _ = pending_response.response_tx.send(msg.to_owned_message());
-    }
-    Ok(())
-}
-/// Send a request and wait for a response with matching sequence number.
-async fn request_on_shared(
-    shared: &Arc<Shared>,
-    msg_type: u8,
-    payload: &[u8],
-    timeout: Duration,
-) -> io::Result<RawMessage> {
-    if timeout.is_zero() {
-        return Err(request_timeout_error());
-    }
-    let deadline = deadline_after(timeout, "request timeout overflowed")?;
-    let seq = shared.next_seq();
-    request_raw_on_shared(shared, msg_type, seq, payload, deadline).await
-}
-
-async fn write_request_frame(
-    shared: &Arc<Shared>,
-    data: &[u8],
-    before_write: impl FnOnce() -> io::Result<()>,
-) -> io::Result<()> {
-    let mut write_guard = RequestWriteGuard::new(Arc::clone(shared));
-    let mut writer = shared.writer.lock().await;
-    before_write()?;
-    write_guard.mark_started();
-    if let Err(error) = writer.write_all(data).await {
-        write_guard.mark_returned();
-        shared.poison_connection();
-        return Err(error);
-    }
-    write_guard.mark_returned();
-    Ok(())
-}
-
-async fn write_request_frame_with_builder(
-    shared: &Arc<Shared>,
-    seq: u32,
-    build_frame: impl FnOnce(u32, &mut Vec<u8>) -> io::Result<()>,
-    before_write: impl FnOnce() -> io::Result<()>,
-) -> io::Result<()> {
-    let mut write_guard = RequestWriteGuard::new(Arc::clone(shared));
-    let frame_builder_guard = shared.frame_builder.lock().await;
-    let mut frame = Vec::new();
-    build_frame(seq, &mut frame)?;
-    let mut writer = shared.writer.lock().await;
-    before_write()?;
-    write_guard.mark_started();
-    let result = writer.write_all(&frame).await;
-    if let Err(error) = result {
-        write_guard.mark_returned();
-        shared.poison_connection();
-        drop(writer);
-        drop(frame);
-        drop(frame_builder_guard);
-        return Err(error);
-    }
-    drop(writer);
-    drop(frame);
-    drop(frame_builder_guard);
-    write_guard.mark_returned();
-    Ok(())
-}
-
-fn encode_request_frame(msg_type: u8, seq: u32, payload: &[u8]) -> io::Result<Vec<u8>> {
-    vsock_proto::encode(msg_type, seq, payload)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e.to_string()))
-}
-
-fn register_pending_response(
-    shared: &Arc<Shared>,
-    seq: u32,
-    build_pending: impl FnOnce(oneshot::Sender<RawMessage>) -> io::Result<PendingResponse>,
-) -> io::Result<oneshot::Receiver<RawMessage>> {
-    // Register under the state lock: `Closed` short-circuits to an
-    // immediate error, and insertion into `pending` is serialised with
-    // the `Connected -> Closed` transition in `close()`. There is no
-    // post-write `is_closed` check because close is observed via the
-    // oneshot receiver becoming `Closed` when `close()` drops the map.
-    let (tx, rx) = oneshot::channel();
-    let mut guard = shared.state.lock().unwrap_or_else(|e| e.into_inner());
-    if let ConnectionState::Connected { pending, .. } = &mut *guard {
-        let pending_response = build_pending(tx)?;
-        pending.insert(seq, pending_response);
-        Ok(rx)
-    } else {
-        Err(io::Error::new(
-            io::ErrorKind::ConnectionReset,
-            "connection closed",
-        ))
-    }
-}
-
-async fn write_registered_request_and_wait(
-    shared: &Arc<Shared>,
-    seq: u32,
-    data: &[u8],
-    deadline: Instant,
-    before_write: impl FnOnce() -> io::Result<()>,
-    rx: oneshot::Receiver<RawMessage>,
-) -> io::Result<RawMessage> {
-    let _pending_guard = PendingRequestGuard::new(Arc::clone(shared), seq);
-
-    // The pending guard removes the pending entry on write failure, timeout,
-    // or cancellation before reader_loop dispatches a response. The write
-    // helper separately poisons the connection if cancellation interrupts an
-    // in-progress frame write.
-    time::timeout_at(deadline, write_request_frame(shared, data, before_write))
-        .await
-        .map_err(|_| request_timeout_error())??;
-
-    await_pending_response(rx, deadline).await
-}
-
-async fn write_registered_request_and_wait_with_frame_builder(
-    shared: &Arc<Shared>,
-    seq: u32,
-    build_frame: impl FnOnce(u32, &mut Vec<u8>) -> io::Result<()>,
-    response_timeout: Duration,
-    before_write: impl FnOnce() -> io::Result<()>,
-    rx: oneshot::Receiver<RawMessage>,
-) -> io::Result<RawMessage> {
-    let _pending_guard = PendingRequestGuard::new(Arc::clone(shared), seq);
-
-    write_request_frame_with_builder(shared, seq, build_frame, before_write).await?;
-
-    let response_deadline = deadline_after(response_timeout, "request timeout overflowed")?;
-    await_pending_response(rx, response_deadline).await
-}
-
-async fn await_pending_response(
-    rx: oneshot::Receiver<RawMessage>,
-    deadline: Instant,
-) -> io::Result<RawMessage> {
-    // `rx` returns `Ok(msg)` when the reader dispatches a response and
-    // `Err(RecvError)` when `close()` drops the `Connected` variant. The
-    // timeout arm is the only other way out.
-    tokio::select! {
-        biased;
-        result = rx => {
-            result.map_err(|_| io::Error::new(
-                io::ErrorKind::ConnectionReset,
-                "connection closed",
-            ))
-        }
-        _ = time::sleep_until(deadline) => {
-            Err(request_timeout_error())
-        }
-    }
-}
-
-fn request_timeout_error() -> io::Error {
-    io::Error::new(io::ErrorKind::TimedOut, "request timeout")
-}
-
-/// Send a request with a pre-allocated sequence number.
-async fn request_raw_on_shared(
-    shared: &Arc<Shared>,
-    msg_type: u8,
-    seq: u32,
-    payload: &[u8],
-    deadline: Instant,
-) -> io::Result<RawMessage> {
-    let data = encode_request_frame(msg_type, seq, payload)?;
-    let rx = register_pending_response(shared, seq, |tx| {
-        Ok(PendingResponse {
-            response_tx: tx,
-            normal_operation: None,
-            normal_terminal_msg_types: &[],
-        })
-    })?;
-
-    write_registered_request_and_wait(shared, seq, &data, deadline, || Ok(()), rx).await
-}
-
-async fn normal_request_on_shared_with_write_observer_frame_builder(
-    shared: &Arc<Shared>,
-    terminal_msg_types: &'static [u8],
-    timeout: Duration,
-    write_observer: FrameWriteObserver,
-    build_frame: impl FnOnce(u32, &mut Vec<u8>) -> io::Result<()>,
-) -> io::Result<RawMessage> {
-    let seq = shared.next_seq();
-    let normal_operation = shared.reserve_normal_operation()?;
-    let rx = register_pending_response(shared, seq, |tx| {
-        Ok(PendingResponse {
-            response_tx: tx,
-            normal_operation: Some(PendingNormalOperation::Owned(normal_operation)),
-            normal_terminal_msg_types: terminal_msg_types,
-        })
-    })?;
-
-    write_registered_request_and_wait_with_frame_builder(
-        shared,
-        seq,
-        build_frame,
-        timeout,
-        || {
-            mark_pending_normal_operation_possible_guest_write(shared, seq, |_| Ok(()))?;
-            write_observer.record_write_start()
-        },
-        rx,
-    )
-    .await
-}
-
-async fn request_on_shared_with_composite_operation_and_observer_frame_builder(
-    shared: &Arc<Shared>,
-    terminal_msg_types: &'static [u8],
-    timeout: Duration,
-    normal_operation: &mut CompositeNormalOperation,
-    write_observer: FrameWriteObserver,
-    build_frame: impl FnOnce(u32, &mut Vec<u8>) -> io::Result<()>,
-) -> io::Result<RawMessage> {
-    let seq = shared.next_seq();
-    let rx = register_pending_response(shared, seq, |tx| {
-        Ok(PendingResponse {
-            response_tx: tx,
-            normal_operation: Some(PendingNormalOperation::Composite(
-                normal_operation.transition_handle()?,
-            )),
-            normal_terminal_msg_types: terminal_msg_types,
-        })
-    })?;
-
-    write_registered_request_and_wait_with_frame_builder(
-        shared,
-        seq,
-        build_frame,
-        timeout,
-        || {
-            mark_pending_normal_operation_possible_guest_write(shared, seq, |_| Ok(()))?;
-            write_observer.record_write_start()
-        },
-        rx,
-    )
-    .await
-}
-
-fn mark_pending_normal_operation_possible_guest_write(
-    shared: &Arc<Shared>,
-    seq: u32,
-    pre_write: impl FnOnce(&mut ConnectionState) -> io::Result<()>,
-) -> io::Result<()> {
-    let mut guard = shared.state.lock().unwrap_or_else(|e| e.into_inner());
-    pre_write(&mut guard)?;
-    match &mut *guard {
-        ConnectionState::Connected { pending, .. } => {
-            let Some(pending_response) = pending.get_mut(&seq) else {
-                return Err(io::Error::new(
-                    io::ErrorKind::ConnectionReset,
-                    "normal request closed before frame write",
-                ));
-            };
-            let Some(normal_operation) = pending_response.normal_operation.as_mut() else {
-                return Err(io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    "normal request missing operation token",
-                ));
-            };
-            mark_pending_normal_operation_possible_guest_write_started(normal_operation)
-        }
-        ConnectionState::Closed => Err(io::Error::new(
-            io::ErrorKind::ConnectionReset,
-            "connection closed",
-        )),
-    }
-}
-
-fn mark_pending_normal_operation_possible_guest_write_started(
-    normal_operation: &mut PendingNormalOperation,
-) -> io::Result<()> {
-    match normal_operation {
-        PendingNormalOperation::Owned(normal_operation) => normal_operation
-            .mark_possible_guest_write_started()
-            .map_err(normal_operation_transition_error),
-        PendingNormalOperation::Composite(normal_operation) => normal_operation
-            .mark_possible_guest_write_started()
-            .map_err(normal_operation_transition_error),
-    }
-}
-
-fn complete_pending_normal_operation(normal_operation: PendingNormalOperation) -> io::Result<()> {
-    match normal_operation {
-        PendingNormalOperation::Owned(normal_operation) => normal_operation
-            .complete()
-            .map_err(normal_operation_transition_error),
-        PendingNormalOperation::Composite(normal_operation) => normal_operation
-            .mark_possible_guest_write_completed()
-            .map_err(normal_operation_transition_error),
-    }
-}
-
-fn normal_operation_rejection_error(error: NormalOperationRejection) -> io::Error {
-    match error {
-        NormalOperationRejection::Fenced => io::Error::new(
-            io::ErrorKind::WouldBlock,
-            "normal operations are currently fenced",
-        ),
-        NormalOperationRejection::NotParkable => io::Error::new(
-            io::ErrorKind::ConnectionReset,
-            "normal operations are not available on this connection",
-        ),
-        NormalOperationRejection::Closed => {
-            io::Error::new(io::ErrorKind::ConnectionReset, "connection closed")
-        }
-    }
-}
-
-fn normal_operation_transition_error(error: NormalOperationTransitionError) -> io::Error {
-    io::Error::new(
-        io::ErrorKind::InvalidData,
-        format!("normal operation transition failed: {error:?}"),
-    )
-}
-
-fn protocol_invalid_data(error: impl ToString) -> io::Error {
-    io::Error::new(io::ErrorKind::InvalidData, error.to_string())
-}
-
-fn lifecycle_error_from_response(response: &RawMessage) -> io::Error {
-    match vsock_proto::decode_error(&response.payload) {
-        Ok(message) => io::Error::other(message.to_owned()),
-        Err(error) => protocol_invalid_data(error),
-    }
-}
-
-fn validate_empty_lifecycle_response(
-    response: &RawMessage,
-    payload_name: &'static str,
-) -> io::Result<()> {
-    vsock_proto::decode_empty_payload(payload_name, &response.payload)
-        .map_err(protocol_invalid_data)
-}
-
-fn deadline_after(timeout: Duration, overflow_message: &'static str) -> io::Result<Instant> {
-    Instant::now()
-        .checked_add(timeout)
-        .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidInput, overflow_message))
-}
-
 impl VsockHost {
-    /// Wait for a guest to connect on the vsock UDS path.
-    ///
-    /// Creates a UDS listener at `{vsock_path}_{port}`, accepts the first
-    /// connection, and performs the ready/ping/pong handshake.
-    ///
-    /// Returns [`io::ErrorKind::InvalidInput`] if `timeout` cannot be
-    /// represented as a Tokio deadline.
-    pub async fn wait_for_connection(vsock_path: &str, timeout: Duration) -> io::Result<Self> {
-        let deadline = deadline_after(timeout, "guest connection timeout overflowed")?;
-        let listener_path = format!("{vsock_path}_{}", vsock_proto::VSOCK_PORT);
-
-        // Clean up stale socket
-        let _ = std::fs::remove_file(&listener_path);
-
-        let listener = UnixListener::bind(&listener_path)?;
-        let mut listener_socket = ListenerSocketGuard {
-            path: Some(listener_path.clone()),
-        };
-
-        let accept_result = time::timeout_at(deadline, listener.accept()).await;
-
-        // Stop accepting and unlink the listener socket before the accepted
-        // stream is handed off. The guard still covers cancellation before
-        // this point.
-        drop(listener);
-        listener_socket.remove();
-
-        let (stream, _) = accept_result.map_err(|_| {
-            io::Error::new(
-                io::ErrorKind::TimedOut,
-                format!("guest connection timeout after {}ms", timeout.as_millis()),
-            )
-        })??;
-
-        Self::from_stream(stream, deadline).await
-    }
-
-    /// Build a `VsockHost` from an already-connected stream.
-    ///
-    /// Performs the handshake on the unsplit stream, then splits it and
-    /// spawns the background reader task.
-    async fn from_stream(stream: UnixStream, deadline: Instant) -> io::Result<Self> {
-        // Handshake on the unsplit stream (reader task not running yet).
-        let (stream, handshake_decoder) = Self::handshake(stream, deadline).await?;
-
-        // Grab the raw fd BEFORE splitting — used by Drop to shutdown the
-        // socket and unblock any pending reads (both our reader_loop and the
-        // remote peer).
-        let fd = {
-            use std::os::unix::io::AsRawFd;
-            stream.as_raw_fd()
-        };
-
-        // Split the stream and spawn the reader task.
-        let (read_half, write_half) = stream.into_split();
-
-        let shared = Arc::new(Shared {
-            writer: tokio::sync::Mutex::new(write_half),
-            frame_builder: tokio::sync::Mutex::new(()),
-            file_write_gate: tokio::sync::Mutex::new(()),
-            fd,
-            seq: AtomicU32::new(2),
-            state: std::sync::Mutex::new(ConnectionState::Connected {
-                pending: HashMap::new(),
-                operations: exec_operation::Operations::new(),
-            }),
-            normal_operations: NormalOperationTracker::new(),
-            close_notify: Notify::new(),
-            #[cfg(test)]
-            exec_output_before_copy_hook: std::sync::Mutex::new(None),
-        });
-
-        let reader_shared = Arc::clone(&shared);
-        let reader = tokio::spawn(reader_loop(read_half, handshake_decoder, reader_shared));
-
-        Ok(Self {
-            shared,
-            file_write_path_locks: file::FileWritePathLocks::default(),
-            _reader: reader,
-            fd,
-        })
-    }
-
-    /// Perform the connection handshake: ready → ping → pong.
-    ///
-    /// Returns the stream and decoder for reuse by the reader task.
-    async fn handshake(
-        mut stream: UnixStream,
-        deadline: Instant,
-    ) -> io::Result<(UnixStream, Decoder)> {
-        let mut decoder = Decoder::new();
-        let mut buf = Box::new([0u8; READ_BUF_SIZE]);
-
-        // Wait for ready
-        Self::read_until_handshake(&mut stream, &mut decoder, &mut buf, deadline, |m| {
-            m.msg_type == MSG_READY
-        })
-        .await?;
-
-        // Send ping with a fixed seq. Shared.seq starts at 2 to avoid collision.
-        let ping_seq: u32 = 1;
-        let ping = vsock_proto::encode(MSG_PING, ping_seq, &[])
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e.to_string()))?;
-        stream.write_all(&ping).await?;
-
-        // Wait for pong with matching seq
-        Self::read_until_handshake(&mut stream, &mut decoder, &mut buf, deadline, |m| {
-            m.msg_type == MSG_PONG && m.seq == ping_seq
-        })
-        .await?;
-
-        Ok((stream, decoder))
-    }
-
-    /// Read messages until one matches the predicate (used during handshake only).
-    async fn read_until_handshake(
-        stream: &mut UnixStream,
-        decoder: &mut Decoder,
-        buf: &mut [u8; READ_BUF_SIZE],
-        deadline: Instant,
-        predicate: impl Fn(&RawMessage) -> bool,
-    ) -> io::Result<RawMessage> {
-        loop {
-            let n = time::timeout_at(deadline, stream.read(buf.as_mut()))
-                .await
-                .map_err(|_| io::Error::new(io::ErrorKind::TimedOut, "handshake timeout"))??;
-            if n == 0 {
-                return Err(io::Error::new(
-                    io::ErrorKind::ConnectionReset,
-                    "connection closed during handshake",
-                ));
-            }
-            let messages = decoder
-                .decode(buf.get(..n).unwrap_or_default())
-                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e.to_string()))?;
-            for msg in messages {
-                if predicate(&msg) {
-                    return Ok(msg);
-                }
-            }
-        }
-    }
-
-    /// Send a request and wait for a response with matching sequence number.
-    async fn request(
-        &self,
-        msg_type: u8,
-        payload: &[u8],
-        timeout: Duration,
-    ) -> io::Result<RawMessage> {
-        request_on_shared(&self.shared, msg_type, payload, timeout).await
-    }
-
-    async fn lifecycle_request(
-        &self,
-        request_type: u8,
-        expected_response_type: u8,
-        response_payload_name: &'static str,
-        timeout: Duration,
-    ) -> io::Result<()> {
-        let response = self.request(request_type, &[], timeout).await?;
-        if response.msg_type == MSG_ERROR {
-            return Err(lifecycle_error_from_response(&response));
-        }
-        if response.msg_type != expected_response_type {
-            return Err(protocol_invalid_data(format!(
-                "unexpected lifecycle response type: expected 0x{expected_response_type:02X}, got 0x{:02X}",
-                response.msg_type,
-            )));
-        }
-        validate_empty_lifecycle_response(&response, response_payload_name)
-    }
-
     /// Try to fence new normal guest operations on this connection.
     ///
     /// The returned guard keeps normal operations fenced until it is dropped.
@@ -1154,37 +288,6 @@ impl VsockHost {
             .try_fence()
             .map(|inner| NormalOperationFence { _inner: inner })
             .map_err(Into::into)
-    }
-
-    /// Ask the guest to quiesce its own operation dispatcher.
-    ///
-    /// This does not fence host-side normal operations. Callers that need a
-    /// no-new-normal-operation boundary must hold a
-    /// [`NormalOperationFence`] before sending this lifecycle request.
-    /// The timeout covers waiting for the shared writer, writing the request,
-    /// and waiting for the response.
-    pub async fn quiesce_operations(&self, timeout: Duration) -> io::Result<()> {
-        self.lifecycle_request(
-            MSG_QUIESCE_OPERATIONS,
-            MSG_OPERATIONS_QUIESCED,
-            "operations_quiesced payload must be empty",
-            timeout,
-        )
-        .await
-    }
-
-    /// Resume guest operations after a failed or aborted quiesce attempt.
-    ///
-    /// The timeout covers waiting for the shared writer, writing the request,
-    /// and waiting for the response.
-    pub async fn resume_operations(&self, timeout: Duration) -> io::Result<()> {
-        self.lifecycle_request(
-            MSG_RESUME_OPERATIONS,
-            MSG_OPERATIONS_RESUMED,
-            "operations_resumed payload must be empty",
-            timeout,
-        )
-        .await
     }
 
     /// Start a request-scoped exec operation using the exec operation protocol.
@@ -1264,9 +367,7 @@ impl VsockHost {
             .try_reserve_and_fence()
             .map_err(|error| FencedExecError::FenceRejected(error.into()))?;
         let mut guard = FencedNormalOperationGuard {
-            normal_operation: Some(CompositeNormalOperation {
-                normal_operation: Some(normal_operation),
-            }),
+            normal_operation: Some(CompositeNormalOperation::from_token(normal_operation)),
             fence: Some(NormalOperationFence { _inner: fence }),
         };
         let normal_operation = guard.normal_operation.as_mut().ok_or_else(|| {
@@ -1293,60 +394,5 @@ impl VsockHost {
         request: ExecStreamRequest<'_>,
     ) -> io::Result<ExecOperationHandle> {
         exec_operation::exec_operation_stream_on_shared(&self.shared, request).await
-    }
-
-    /// Request graceful shutdown from guest.
-    ///
-    /// The timeout covers waiting for the shared writer, writing the request,
-    /// and waiting for the acknowledgement.
-    ///
-    /// # Errors
-    ///
-    /// Returns the request error if the guest cannot be reached or does not
-    /// acknowledge before the deadline. An unexpected response type or a
-    /// non-empty acknowledgement returns [`io::ErrorKind::InvalidData`].
-    pub async fn shutdown(&self, timeout: Duration) -> io::Result<()> {
-        self.lifecycle_request(
-            MSG_SHUTDOWN,
-            MSG_SHUTDOWN_ACK,
-            "shutdown_ack payload must be empty",
-            timeout,
-        )
-        .await
-    }
-}
-
-#[cfg(test)]
-impl VsockHost {
-    /// Test-only: deterministically await the `Connected → Closed` transition
-    /// without relying on a wall-clock sleep. Subscribes to the same
-    /// `close_notify` signal that [`Shared::close`] fires on exit, and re-checks
-    /// state under the same lock that `close` holds, so no transition is
-    /// missed.
-    async fn wait_until_closed(&self, timeout: Duration) -> io::Result<()> {
-        let deadline = deadline_after(timeout, "wait_until_closed timeout overflowed")?;
-        loop {
-            let notified = self.shared.close_notify.notified();
-            tokio::pin!(notified);
-            notified.as_mut().enable();
-
-            if matches!(
-                &*self.shared.state.lock().unwrap_or_else(|e| e.into_inner()),
-                ConnectionState::Closed
-            ) {
-                return Ok(());
-            }
-
-            tokio::select! {
-                biased;
-                _ = notified => {}
-                _ = tokio::time::sleep_until(deadline) => {
-                    return Err(io::Error::new(
-                        io::ErrorKind::TimedOut,
-                        "wait_until_closed: reader did not transition to Closed in time",
-                    ));
-                }
-            }
-        }
     }
 }
